new commit

+++ Waka/Waka/paddle.h
@@ -9,8 +9,8 @@
 class Paddle final : public Propulsion 
 {
 private:
-    int knots = 0;
-    int do_getKnots() const noexcept override;
+    double knots = 0;
+    double do_getKnots() const noexcept override;
 };
 
 #endif
