+++ Waka/Waka/propulsion.h
@@ -10,11 +10,11 @@ class Propulsion
 {
 public:
     virtual ~Propulsion() = default;
-    int getKnots() const noexcept;
+    double getKnots() const noexcept;
     
 protected:
     //pure virtual function
-    virtual int do_getKnots() const noexcept = 0;
+    virtual double do_getKnots() const noexcept = 0;
 };
 
 #endif
